Remove Newsesque
We no longer need this class following
f0844e4e54a62f3226aeb5793baa57421557972d as only `NewsArticle` inherits
from `Newsesque`.
