+++ app/models/announcement.rb
@@ -32,6 +32,6 @@ class Announcement < Edition
   end
 end
 
-require_relative "newsesque"
+require_relative "news_article"
 require_relative "speech"
 require_relative "fatality_notice"
